+++ src/headers/app.helper.h
@@ -1,8 +1,8 @@
 #ifndef MY_PRINTF_APP_HELPER_H
 #define MY_PRINTF_APP_HELPER_H
 
-extern long int    stringLengthHelper(string s);
-extern void	    charPrintHelper(const char c);
+extern size_t    stringLengthHelper(string s);
+extern void	   charPrintHelper(const char c);
 extern void        stringPrintHelper(const string s);
 extern void        integerPrintHelper(int n);
